+++ Complexity/ComplexityWeb/DTO/TopicInfoDTO.cs
@@ -1,7 +1,7 @@
 using System;
 using Newtonsoft.Json;
 
-namespace Application
+namespace ComplexityWebApi.DTO
 {
     public class TopicInfoDTO
     {
