+++ Complexity/ComplexityWeb/Program.cs
@@ -1,8 +1,7 @@
-﻿using ComplexityWebApi;
-using Microsoft.AspNetCore;
+﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace Complexity
+namespace ComplexityWebApi
 {
 #pragma warning disable CS1591
     public class Program
