+++ Complexity/Domain/Domain.csproj
@@ -6,7 +6,6 @@
 
   <ItemGroup>
     <PackageReference Include="Scriban" Version="2.0.0" />
-    <PackageReference Include="StringTemplate4" Version="4.0.8" />
   </ItemGroup>
 
   <ItemGroup>
