+++ Complexity/Infrastructure/DDD/EntityAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Infrastructure.DDD
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Interface)]
+    public class EntityAttribute : Attribute
+    {
+    }
+}
