+++ Complexity/Infrastructure/DDD/ValueAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Infrastructure.DDD
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct)]
+    public class ValueAttribute : Attribute
+    {
+    }
+}
