+++ Complexity/Infrastructure/Result/Result.cs
@@ -6,52 +6,6 @@ using System.Threading.Tasks;
 
 namespace Infrastructure.Result
 {
-    internal class ResultCommonLogic<TError>
-    {
-        [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private readonly TError error;
-
-        [DebuggerStepThrough]
-        internal ResultCommonLogic(bool isFailure, TError error)
-        {
-            if (isFailure)
-            {
-                if (error == null)
-                    throw new ArgumentNullException(nameof(error), ResultMessages.ErrorObjectIsNotProvidedForFailure);
-            }
-            else
-            {
-                if (error != null)
-                    throw new ArgumentException(ResultMessages.ErrorObjectIsProvidedForSuccess, nameof(error));
-            }
-
-            IsFailure = isFailure;
-            this.error = error;
-        }
-
-        public bool IsFailure { get; }
-        public bool IsSuccess => !IsFailure;
-
-        public TError Error
-        {
-            [DebuggerStepThrough]
-            get
-            {
-                if (IsSuccess)
-                    throw new ResultSuccessException();
-
-                return error;
-            }
-        }
-
-        public void GetObjectData(SerializationInfo info, StreamingContext context)
-        {
-            info.AddValue("IsFailure", IsFailure);
-            info.AddValue("IsSuccess", IsSuccess);
-            if (IsFailure) info.AddValue("Error", Error);
-        }
-    }
-
     internal sealed class ResultCommonLogic : ResultCommonLogic<string>
     {
         private ResultCommonLogic(bool isFailure, string error) : base(isFailure, error)
@@ -74,27 +28,6 @@ namespace Infrastructure.Result
 
             return new ResultCommonLogic(isFailure, error);
         }
-    }
-
-    internal static class ResultMessages
-    {
-        public static readonly string ErrorIsInaccessibleForSuccess =
-            "You attempted to access the Error property for a successful result. A successful result has no Error.";
-
-        public static readonly string ValueIsInaccessibleForFailure =
-            "You attempted to access the Value property for a failed result. A failed result has no Value.";
-
-        public static readonly string ErrorObjectIsNotProvidedForFailure =
-            "You attempted to create a failure result, which must have an error, but a null error object was passed to the constructor.";
-
-        public static readonly string ErrorObjectIsProvidedForSuccess =
-            "You attempted to create a success result, which cannot have an error, but a non-null error object was passed to the constructor.";
-
-        public static readonly string ErrorMessageIsNotProvidedForFailure =
-            "You attempted to create a failure result, which must have an error, but a null or empty string was passed to the constructor.";
-
-        public static readonly string ErrorMessageIsProvidedForSuccess =
-            "You attempted to create a success result, which cannot have an error, but a non-null string was passed to the constructor.";
     }
 
     public struct Result : IResult, ISerializable
