+++ Complexity/Infrastructure/Result/ResultFailureException.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Infrastructure.Result
+{
+    public class ResultFailureException : Exception
+    {
+        internal ResultFailureException(string error) : base(ResultMessages.ValueIsInaccessibleForFailure)
+        {
+            Error = error;
+        }
+
+        public string Error { get; }
+    }
+}
