+++ Complexity/Infrastructure/Result/ResultSuccessException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Infrastructure.Result
+{
+    public class ResultSuccessException : Exception
+    {
+        internal ResultSuccessException() : base(ResultMessages.ErrorIsInaccessibleForSuccess)
+        {
+        }
+    }
+}
