+++ Complexity/Tests/Tests.csproj
@@ -24,8 +24,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Application\Application.csproj" />
-    <ProjectReference Include="..\ComplexityWeb\ComplexityWebApi.csproj" />
-    <ProjectReference Include="..\DataBase\DataBase.csproj" />
   </ItemGroup>
 
 </Project>
